tests/wpt/web-platform-tests/service-workers/service-worker/fetch-event-respond-with-readable-stream.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>respondWith with a response built from a ReadableStream</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
'use strict';

const WORKER =
  'resources/fetch-event-respond-with-readable-stream-worker.js';
const SCOPE =
  'resources/fetch-event-respond-with-readable-stream-iframe.html';
// Called by the iframe when done.
var done;
var done_was_called = new Promise(resolve => done = resolve);

promise_test(t => {
    return service_worker_unregister_and_register(t, WORKER, SCOPE)
      .then(reg => {
           add_completion_callback(() => reg.unregister());
           return wait_for_state(t, reg.installing, 'activated');
         })
      .then(() => with_iframe(SCOPE))
      .then(iframe => {
          t.add_cleanup(() => iframe.remove())
        })
      .then(() => done_was_called)
      .then(result => assert_equals(result, 'PASS'));
  }, 'Respond with a Response built from a ReadableStream');
</script>
